Adds PlayersRepository and renames GamesRepository
